Following changes have been implemented. To move AshFall related class to ashfall directory. To update javadoc. To implement disaster interfaces. To create test for AshFallParserTest. To modify test for MarineParserTest.
